Use RTAIL/LTAIL, Illumina first: 1/9 tests fails
Reverse the order of trimming: trim Illumina | trim TSO (was: trim TSO
| trim Illumina, which was in the order from most to least frequent
adapters). This change is done to prevent r1_ins_1_28_tso_full_g5_rev:
from failing, which it did. The number of failed tests is reduced from
2 to 1:

r1_ins_1_40_ilmn_13_fwd: as expected, 20 nt Illumina adapter does not
reach all the way to the end of the read, so both r1 and r2 are not
trimmed.